Fixes the bias towards getting larger when resizing.
Passing 'true' to the pixelsToRows and pixelsToColumns functions causes
these functions to use Math.ceil to calculate how many rows or columns
the item should occupy. This biases towards getting larger and makes
resizing experience undesired. Similarly, 'false' to these functions
uses Math.floor. This isn't as important for us since we do not allow
resize handles on the left side of charts, but if we did the user would
have similarly poor experiences.

This also has the effect of fixing a bug when resizing off the grid.
Before, if you resized an item all the way to the right of the grid, the
item would move to the left and push other items down. Now, you cannot
resize past the edge of the grid.